Add multilanguage support
Articles now have a language field that describes the
language used in the article. It also excepts empty value which means
the article is language agnostic.

In article listings the list can be filtered with language optional
language parameter. CMS plugins also have a language filter field so
that you can either filter the listing with specific language or have
a language agnostic listing.